refactor: use `BackgroundThreads` from `common`
fixes #1250
