+++ google/cloud/spanner/connection_options.cc
@@ -13,8 +13,8 @@
 // limitations under the License.
 
 #include "google/cloud/spanner/connection_options.h"
-#include "google/cloud/spanner/internal/background_threads_impl.h"
 #include "google/cloud/spanner/internal/compiler_info.h"
+#include "google/cloud/internal/background_threads_impl.h"
 #include "google/cloud/internal/getenv.h"
 #include "google/cloud/log.h"
 #include <sstream>
@@ -39,7 +39,7 @@ ConnectionOptions::ConnectionOptions(
       user_agent_prefix_(internal::BaseUserAgentPrefix()),
       background_threads_factory_([] {
         return google::cloud::internal::make_unique<
-            internal::AutomaticallyCreatedBackgroundThreads>();
+            google::cloud::internal::AutomaticallyCreatedBackgroundThreads>();
       }) {
   auto tracing =
       google::cloud::internal::GetEnv("GOOGLE_CLOUD_CPP_ENABLE_TRACING");
@@ -85,7 +85,7 @@ ConnectionOptions& ConnectionOptions::DisableBackgroundThreads(
     google::cloud::grpc_utils::CompletionQueue const& cq) {
   background_threads_factory_ = [cq] {
     return google::cloud::internal::make_unique<
-        internal::CustomerSuppliedBackgroundThreads>(cq);
+        google::cloud::internal::CustomerSuppliedBackgroundThreads>(cq);
   };
   return *this;
 }
