+++ google/cloud/spanner/spanner_client_unit_tests.bzl
@@ -27,7 +27,6 @@ spanner_client_unit_tests = [
     "instance_admin_connection_test.cc",
     "instance_test.cc",
     "internal/api_client_header_test.cc",
-    "internal/background_threads_test.cc",
     "internal/build_info_test.cc",
     "internal/compiler_info_test.cc",
     "internal/connection_impl_test.cc",
